Collisions with terrain now work
The problem was that the sprite's position wasn't aligning with its x and y values. That is because the sprite is drawn in the middle of the screen at all times so the sprite location is always at (320, 320). This is ultimately the problem of the sprite being used to draw and being used as a collision rect. Therefore, I made an additional drawSprite for playerCity.