Properly separated the use of worldmaps/terrain tiles and cities
Basically, the game itself was doing too much of the cities work. I wanted the city to actually have control of its own stuff. like moveSpeed and calculating that. I wanted the worldmap and city to interact via the game loop, but not use one another within themselves. This has fixed that by having the execute method be the new place where all the terrain interaction + movement behavior occurs.