Implment proper border handling
There was crashing before because the x and y of the city sprite were such that it would request terrain tiles that didn't exist and therefore created errors. Now the x and y are trimmed to the max or the min dependin on where the city is. Also, movement is restricted to the map area.